+++ development/http_client/dev_http_client.py
@@ -17,10 +17,3 @@ rpc_client = pyjsonrpc.HttpClient("http://localhost:8080", gzipped = False)
 # Example with *call*
 print repr(rpc_client.call("add", 1, 2))
 
-# # Create very large string
-# s = u"We are the champions! " * (47662 * 15)  # 15 MiB
-#
-# # Example
-# print rpc_client.add(u"Long string: ", s)[:100]
-#
-